Ahh screw this python :(
