+++ .gitignore
@@ -7,7 +7,7 @@ __pycache__/
 *.so
 
 #Secrets
-secrets.py
+*secrets*
 
 # Distribution / packaging
 .Python
