+++ ToIssue/webhooks/logic.py
@@ -1,6 +1,4 @@
-# Local imports
-from .todoist_help_functions import *
-from .secrets import *
+from secrets import *
 
 # Outbound imports
 import os
@@ -9,9 +7,27 @@ from todoist.api import TodoistAPI
 from github import Github
 
 
-gitApi = Github(GITHUB_API_TOKEN)
-api = TodoistAPI(TODOIST_API_TOKEN)
-api.sync()
+def find_project_by_name(name):
+    """ Help function to find specific projects ID by given name
+
+    Args:
+        name ([str]): [Project name]
+    """
+    api = TodoistAPI(TODOIST_API_TOKEN)
+    api.sync()
+    projects = api.state['projects']
+    for project in projects:
+        if project['name'] == name:
+            return(project['id'])
+
+def find_task_by_name(content):
+    api = TodoistAPI(TODOIST_API_TOKEN)
+    api.sync()
+    items = api.state['items']
+    for item in items:
+        if item['content'] == content:
+            return(item['id'])
+
 
 # TODO: Add if for the object added to the todolist, while right now it's all added no matter 
 # if is in the list or nah
@@ -22,6 +38,11 @@ def getAllMyIssuesFromProject(repo, username):
         repo ([Repository]): [github repository we want to get the issues from]
         username ([str]): [username that we want all the issues to be added to our "Todoist App"]
     """
+
+    gitApi = Github(GITHUB_API_TOKEN)
+    api = TodoistAPI(TODOIST_API_TOKEN)
+    api.sync()
+
     repo = gitApi.get_repo(f"{username}/{repo}")
     todoProject = api.projects.get_by_id(find_project_by_name(repo.full_name))
     issues = repo.get_issues(state="open")
@@ -33,8 +54,6 @@ def getAllMyIssuesFromProject(repo, username):
                     api.commit()
                 else:
                     pass
-            else:
-                pass
     else:
         todoProject = api.projects.add(f"{repo.full_name}")
         for issue in issues:
@@ -44,11 +63,29 @@ def getAllMyIssuesFromProject(repo, username):
                     api.commit()
                 else:
                     pass
-            else:
-                pass
 
-def addNewIssueToTodoist(repo, username, content):
-    return "wroking"
-    
-def checkingifIssue(data):
-    return None
+def addNewIssueToTodoist(repoFullName, username, issueTitle):
+    api = TodoistAPI(TODOIST_API_TOKEN)
+    api.sync()
+    todoProject = api.projects.get_by_id(find_project_by_name(repoFullName))
+    print(todoProject)
+    if(todoProject != None):
+        if username == 'LSWarss':
+            if api.items.get_by_id(find_task_by_name(issueTitle) != None):
+                task = api.items.add(issueTitle, project_id=todoProject['id'])
+                print(task)
+                api.commit()
+        else:
+            pass
+    else:
+        todoProject = api.projects.add(repoFullName)
+        if username == 'LSWarss':
+            if api.items.get_by_id(find_task_by_name(issueTitle) != None):
+                task = api.items.add(issueTitle, project_id=todoProject['id'])
+                api.commit()
+        else:
+            pass
+
+
+addNewIssueToTodoist("LSWarss/ToIssue","LSWarss","Testing issue 10")
+getAllMyIssuesFromProject("LSWarss/ToIssue","LSWarss")
